iterating on the refactor
