+++ anvilfs/anvilfs.py
@@ -1,43 +1,13 @@
 from .base import BaseAnVILFolder
 from .namespace import Namespace
 from .workspace import Workspace
+from .workloadidentitycredentials import WorkloadIdentityCredentials
 
-from fs.base import FS
-from fs.errors import DirectoryExpected, ResourceNotFound, FileExpected
-
-import datetime
-from google.auth import credentials
-import json
 import firecloud.api as fapi
 from google.auth.transport.requests import AuthorizedSession
 
-class WorkloadIdentityCredentials(credentials.Scoped, credentials.Credentials):
-    def __init__(self, scopes):
-        super(WorkloadIdentityCredentials, self).__init__()
-        print(f"Init with scopes={scopes}")
-        self._scopes = scopes
-
-    def with_scopes(self, scopes):
-        return WorkloadIdentityCredentials(scopes=scopes)
-
-    @property
-    def requires_scopes(self):
-        return False
-
-    def refresh(self, request):
-        print(f"Refresh with scopes={self._scopes}")
-        url = 'http://metadata.google.internal/computeMetadata/v1/instance/service-accounts/default/token'
-        if self._scopes:
-            url += '?scopes=' + ','.join(self._scopes)
-        response = request(url=url, method="GET", headers={
-                           'Metadata-Flavor': 'Google'})
-        if response.status == 200:
-            response_json = json.loads(response.data)
-        else:
-            raise RuntimeError('bad status from metadata server')
-        self.token = response_json['access_token']
-        self.expiry = datetime.datetime.utcnow(
-        ) + datetime.timedelta(seconds=response_json['expires_in'])
+from fs.base import FS
+from fs.errors import DirectoryExpected, ResourceNotFound, FileExpected
 
 
 class AnVILFS(FS):
@@ -51,7 +21,7 @@ class AnVILFS(FS):
 #        fapi.list_workspaces()
         # /hax
         self.namespace = Namespace(namespace)
-        self.workspace = self.namespace.fetch_workspace(workspace)
+        self.workspace = self.namespace.fetch_workspace(workspace, fapi)
         self.rootobj = self.workspace  # leaving the option to make namespace root
 
     def getinfo(self, path, namespaces=None):
