+++ anvilfs/namespace.py
@@ -5,8 +5,8 @@ class Namespace(BaseAnVILFolder):
     def __init__(self, namespace_name):
         super().__init__(namespace_name)
 
-    def fetch_workspace(self, workspace_name):
-        ws = Workspace(self, workspace_name)
+    def fetch_workspace(self, workspace_name, fapi):
+        ws = Workspace(self, workspace_name, fapi)
         self[ws.name] = ws
         return ws
